BUGFIX: Atma Weapon
Turns out the Atma Weapon artifact was completely nonfunctional o_O. I thought I tested that....

+++ dnethack-3.4.3/src/weapon.c
@@ -263,8 +263,8 @@ struct monst *mon;
 			){
 				tmp += u.ulevel;
 				tmp *= Upolyd ?
-						u.mh/u.mhmax  :
-						u.uhp/u.uhpmax;
+						((float)u.mh)/u.mhmax  :
+						((float)u.uhp)/u.uhpmax;
 			}
 			break;
 		case DOUBLE_LIGHTSABER: 
@@ -317,8 +317,8 @@ struct monst *mon;
 			){
 				tmp += u.ulevel;
 				tmp *= Upolyd ?
-						u.mh/u.mhmax  :
-						u.uhp/u.uhpmax;
+						((float)u.mh)/u.mhmax  :
+						((float)u.uhp)/u.uhpmax;
 			}
 			break;
 		case DOUBLE_LIGHTSABER:
